move health main to duckbot.health
